test: fix repotool-export-cvs prerequisite check
Both cvs and cvs-fast-export are required.

+++ test/repotool-export-cvs.sh
@@ -1,6 +1,7 @@
 ## Test repotool export of CVS repo
 
 command -v cvs >/dev/null 2>&1 || { echo "    Skipped, cvs missing."; exit 0; }
+command -v cvs-fast-export >/dev/null 2>&1 || { echo "    Skipped, cvs-fast-export missing."; exit 0; }
 
 trap "rm -rf /tmp/test-repo$$ /tmp/target$$ /tmp/out$$" 0 12 2 15
